requirements.txt
# TurtleBot3 Smart Automation - Python Dependencies
# ===================================================

# Core dependencies
pyyaml>=6.0
psutil>=5.9.0

# Computer Vision
ultralytics>=8.0.0
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0

# ROS2 Python libraries (installed via apt, but listed for reference)
# ros-humble-rclpy
# ros-humble-std-msgs
# ros-humble-geometry-msgs
# ros-humble-sensor-msgs
# ros-humble-nav-msgs

# Optional dependencies for enhanced features
pillow>=10.0.0
matplotlib>=3.7.0
scipy>=1.11.0
